Update Metrics Payload Structure
After some discussion with @aliok (who did lots of research around the mobile metrics and built the PoC), We've decided to go with the approach of storing the metrics data in a JSONB column called `data`. The proposed structure here would be the simplest to parse into a struct (in Golang), to strip out undesired fields and do validation.